@@ -1,7 +1,8 @@
 {
   "name": "o-ngx-tree",
-  "version": "4.0.0",
-  "description": "Angular component for visualizing data that can be naturally represented as a tree. Forked from valor-software/ng2-tree",
+  "version": "8.0.0-SNAPSHOT-0",
+  "description":
+    "Angular component for visualizing data that can be naturally represented as a tree. Forked from valor-software/ng2-tree",
   "main": "index.js",
   "license": "MIT",
   "author": "Imatia S.L.",
@@ -28,7 +29,8 @@
     "recursive"
   ],
   "scripts": {
-    "pre:publish": "npm run clean && npm run test && npm run lint && npm run compile && mkdir -p dist && cp -R src/styles.css README.md media dist",
+    "pre:publish":
+      "npm run clean && npm run test && npm run lint && npm run compile && mkdir -p dist && cp -R src/styles.css README.md media dist",
     "post:publish": "npm run build:demo && gh-pages -d dist-demo",
     "start": "ng serve --prod",
     "build:demo": "ng build",
@@ -55,27 +57,26 @@
     }
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.6.6",
-    "@angular/cli": "6.0.3",
-    "@angular/common": "6.0.3",
-    "@angular/compiler": "6.0.3",
-    "@angular/compiler-cli": "6.0.3",
-    "@angular/core": "6.0.3",
-    "@angular/forms": "6.0.3",
-    "@angular/http": "6.0.3",
-    "@angular/language-service": "6.0.3",
-    "@angular/platform-browser": "6.0.3",
-    "@angular/platform-browser-dynamic": "6.0.3",
-    "@angular/router": "6.0.3",
+    "@angular-devkit/build-angular": "~0.803.22",
+    "@angular/cli": "~8.3.21",
+    "@angular/common": "~8.2.14",
+    "@angular/compiler": "~8.2.14",
+    "@angular/compiler-cli": "~8.2.14",
+    "@angular/core": "~8.2.14",
+    "@angular/forms": "~8.2.14",
+    "@angular/language-service": "~8.2.14",
+    "@angular/platform-browser": "~8.2.14",
+    "@angular/platform-browser-dynamic": "~8.2.14",
+    "@angular/router": "~8.2.14",
     "@types/jasmine": "2.8.2",
-    "@types/node": "8.0.53",
+    "@types/node": "~8.9.4",
     "alertifyjs": "1.10.0",
-    "codelyzer": "4.0.1",
+    "codelyzer": "5.2.1",
     "conventional-changelog": "1.1.7",
     "conventional-changelog-cli": "1.3.5",
     "conventional-github-releaser": "2.0.0",
     "copyfiles": "1.2.0",
-    "core-js": "2.5.1",
+    "core-js": "^3.6.4",
     "font-awesome": "4.7.0",
     "gh-pages": "1.1.0",
     "husky": "0.14.3",
@@ -95,7 +96,7 @@
     "pretty-quick": "1.4.1",
     "protractor": "5.2.0",
     "rimraf": "2.6.2",
-    "rxjs": "6.2.0",
+    "rxjs": "~6.4.0",
     "rxjs-compat": "^6.2.0",
     "rxjs-tslint": "^0.1.4",
     "shelljs": "0.7.8",
@@ -103,9 +104,9 @@
     "ts-node": "3.3.0",
     "tslint": "5.8.0",
     "tslint-config-valorsoft": "2.1.1",
-    "typescript": ">=2.7.2 <2.8",
+    "typescript": "~3.5.3",
     "uuid": "3.1.0",
     "webpack": "^4.10.2",
-    "zone.js": "0.8.26"
+    "zone.js": "0.9.1"
   }
 }
